+++ packages/riverpod_lint_flutter_test/test/lints/provider_parameters.dart
@@ -268,6 +268,5 @@ class IndirectEqual with Equatable {
 class ClassicFreezed with _$ClassicFreezed {
   const ClassicFreezed(this.value);
 
-  @override
   final int value;
 }