+++ packages/flutter_riverpod/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.4.3 - 2026-09-04
+### Dependency changes
+
+- `riverpod` upgraded to `3.4.3`
+
 ## 3.4.2 - 2026-07-28
 
 Fix a different source of `markNeedsBuild` error. Those are tricky!